+++ controllers/courses.js
@@ -9,7 +9,7 @@ const advancedResults = require('../middleware/advanceResults')
 // @routes  GET /api/v1/bootcamps/:bootcampId/courses
 // @access  Public
 exports.getCourses = asyncHandler(async (req, res, next) => {
-    if (req.params.bootcampId) {        
+    if (req.params.bootcampId) {
         const courses = await Course.find({
             bootcamp: req.params.bootcampId
         })
@@ -48,8 +48,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
         select: 'name description'
     })
 
-    if(!course)
-    {
+    if (!course) {
         return next(new ErrorResponse(`No Course with the id of ${req.params.id}`), 404)
     }
 
@@ -65,12 +64,19 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 exports.createCourse = asyncHandler(async (req, res, next) => {
 
     req.body.bootcamp = req.params.bootcampId
-    
+
+    req.body.user = req.user.id
+
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
-    if(!bootcamp)
-    {
+    if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404)
+    }
+
+    // Make sure user is bootcamp user
+    if (bootcamp.user.toString() !== req.user.id && req.user.role != 'admin') {
+        console.log('\n78 hello \n');
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`, 401))
     }
 
     const course = await Course.create(req.body)
@@ -88,9 +94,14 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 
     let course = await Course.findById(req.params.id);
 
-    if(!course)
-    {
+    if (!course) {
         return next(new ErrorResponse(`No Course with the id of ${req.params.bootcampId}`), 404)
+    }
+
+    // Make sure user is bootcamp user
+    if (course.user.toString() !== req.user.id && req.user.role != 'admin') {
+        console.log('\n78 hello \n');
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update a course ${course._id}`, 401))
     }
 
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -110,13 +121,18 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 
     const course = await Course.findById(req.params.id);
 
-    if(!course)
-    {
+    if (!course) {
         return next(new ErrorResponse(`No Course with the id of ${req.params.id}`), 404)
     }
 
+    // Make sure user is bootcamp user
+    if (course.user.toString() !== req.user.id && req.user.role != 'admin') {
+        console.log('\n78 hello \n');
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete a course ${course._id}`, 401))
+    }
+
     await course.deleteOne();
-   
+
     res.status(200).json({
         success: true,
         data: {}
